speedy/requirements.txt
Django==4.1.1
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
Pillow==9.2.0
drf-yasg==1.21.4
psycopg2==2.9.3